Abstracted all capture methods to its own module
