+++ src/capture_method/ForceFullContentRenderingCaptureMethod.py
@@ -1,0 +1,5 @@
+from capture_method.BitBltCaptureMethod import BitBltCaptureMethod
+
+
+class ForceFullContentRenderingCaptureMethod(BitBltCaptureMethod):  # pylint: disable=too-few-public-methods
+    _render_full_content = True
