+++ src/utils.py
@@ -1,12 +1,17 @@
 import asyncio
+import ctypes
+import ctypes.wintypes
 import os
 import sys
 from collections.abc import Callable
 from platform import version
-from typing import Any, Optional, Union
+from typing import Any, Optional, Union, cast
 
 import cv2
 from typing_extensions import TypeGuard
+from win32 import win32gui
+
+DWMWA_EXTENDED_FRAME_BOUNDS = 9
 
 
 def decimal(value: Union[int, float]):
@@ -27,6 +32,22 @@ def is_digit(value: Optional[Union[str, int]]):
 
 def is_valid_image(image: Optional[cv2.Mat]) -> TypeGuard[cv2.Mat]:
     return image is not None and bool(image.size)
+
+
+def get_window_bounds(hwnd: int):
+    extended_frame_bounds = ctypes.wintypes.RECT()
+    ctypes.windll.dwmapi.DwmGetWindowAttribute(
+        hwnd,
+        DWMWA_EXTENDED_FRAME_BOUNDS,
+        ctypes.byref(extended_frame_bounds),
+        ctypes.sizeof(extended_frame_bounds))
+
+    window_rect = win32gui.GetWindowRect(hwnd)
+    window_left_bounds = cast(int, extended_frame_bounds.left) - window_rect[0]
+    window_top_bounds = cast(int, extended_frame_bounds.top) - window_rect[1]
+    window_width = cast(int, extended_frame_bounds.right) - cast(int, extended_frame_bounds.left)
+    window_height = cast(int, extended_frame_bounds.bottom) - cast(int, extended_frame_bounds.top)
+    return window_left_bounds, window_top_bounds, window_width, window_height
 
 
 def fire_and_forget(func: Callable[..., None]):
